requirements.txt
Pillow<9.5.0
Requests==2.31.0
streamlit==1.24.1
streamlit_lottie==0.0.5
pandas<1.2.0
